+++ Changelog.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 ### Added
+- Support `--recurrent-only` option for `fina list` command.
 ### Changed
 ### Removed
 ### Fixed