+++ Changelog.md
@@ -5,11 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 ### Added
-- Support `--recurrent-only` option for `fina list` command.
 ### Changed
-- The upstream `financeager` dependency is required as v1.0.1 or higher.
 ### Removed
 ### Fixed
+
+## [v0.4.0.0] - 2022-02-04
+### Added
+- Support `--recurrent-only` option for `fina list` command.
+### Changed
+- The upstream `financeager` dependency is required as v1.0.0 or higher.
 
 ## [v0.3.4.1 - 2021-12-30]
 ### Fixed